refactor: created the NewGiftForm component and implemented id on the OwnerList component
